+++ packages/patrol/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `setMockLocation` method. (#2547)
 - Bump `custom_lint` to `0.7.0` and `leancode_lint` to `14.3.0`. (#2574)
 - Upgrade `http4k` to 5.47.0.0. (#2559)
+- Make `PatrolBinding`'s `testResults` public. (#2362)
 
 ## 3.14.1